backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.22
pikepdf==8.12.0
pypdf==4.1.0
pdf2image==1.17.0
Pillow==12.1.1
reportlab==4.1.0
aiofiles==23.2.1
pytesseract==0.3.13
weasyprint
slowapi==0.1.9
